eliminar el campo valor de Subservicio, y revajar el costo del subservicio asociado a la empresa del monto de la empresa

+++ src/Entity/Telecomunicaciones/EmpresaTipoPaga.php
@@ -24,7 +24,7 @@ class EmpresaTipoPaga
     private $tipo;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $saldo;
 
@@ -51,12 +51,12 @@ class EmpresaTipoPaga
         return $this;
     }
 
-    public function getSaldo(): ?int
+    public function getSaldo(): ?float
     {
         return $this->saldo;
     }
 
-    public function setSaldo(int $saldo): self
+    public function setSaldo(float $saldo): self
     {
         $this->saldo = $saldo;
